+++ src/main/java/com/demo/eventi/service/impl/PrenotazioneService.java
@@ -2,10 +2,16 @@ package com.demo.eventi.service.impl;
 
 import java.util.Optional;
 
+import org.springframework.stereotype.Service;
+
 import com.demo.eventi.model.Prenotazione;
 import com.demo.eventi.repository.PrenotazioneRepository;
 import com.demo.eventi.service.IPrenotazioneService;
 
+import jakarta.transaction.Transactional;
+
+@Service
+@Transactional
 public class PrenotazioneService implements IPrenotazioneService {
 	
 private final PrenotazioneRepository prenotazioneRepository;
@@ -16,20 +22,26 @@ private final PrenotazioneRepository prenotazioneRepository;
 
 	@Override
 	public Prenotazione creaPrenotazione(Prenotazione prenotazione) {
-		// vedere se servono controlli
 		return prenotazioneRepository.save(prenotazione);
 	}
 	
 	@Override
-	public Prenotazione aggiornaPrenotazione(Prenotazione prenotazione) {
-		return prenotazioneRepository.save(prenotazione);
+	public Prenotazione aggiornaPrenotazione(Prenotazione prenotazione) throws Exception {
+	    if (!prenotazioneRepository.existsById(prenotazione.getIdPrenotazione())) {
+	        throw new Exception("Prenotazione non trovata per ID: " + prenotazione.getIdPrenotazione());
+	    }
+	    return prenotazioneRepository.save(prenotazione);
 	}
 
+
 	@Override
-	public void eliminaPrenotazione(Long id) {
-		prenotazioneRepository.deleteById(id);
-		
+	public void eliminaPrenotazione(Long id) throws Exception {
+	    if (!prenotazioneRepository.existsById(id)) {
+	        throw new Exception("Prenotazione non trovata per ID: " + id);
+	    }
+	    prenotazioneRepository.deleteById(id);
 	}
+
 
 	@Override
 	public Optional<Prenotazione> trovaPerId(Long id) {
